ORC-1502: Upgrade Maven to 3.9.4
### What changes were proposed in this pull request?

This PR aims to upgrade Maven to 3.9.4.

### Why are the changes needed?

Apache Maven community starts to test Java 20 officially. Although this is only about a test coverage, we had better adopt it because 3.8.x is not tested officially by Apache Maven community.
- https://issues.apache.org/jira/browse/MNG-7743
- https://github.com/apache/maven/pull/1065

Here are the release notes.
- https://maven.apache.org/docs/3.9.0/release-notes.html
- https://maven.apache.org/docs/3.9.1/release-notes.html
- https://maven.apache.org/docs/3.9.2/release-notes.html
- https://maven.apache.org/docs/3.9.3/release-notes.html
- https://maven.apache.org/docs/3.9.4/release-notes.html

### How was this patch tested?

Pass the CIs with Java 21 EA build

Closes #1612 from dongjoon-hyun/ORC-1502.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>