ORC-1504: Add lower bound check for get() in DynamicIntArray
### What changes were proposed in this pull request?
The PR proposes adding check for lower bound for index passed to get() in DynamicIntArray

### Why are the changes needed?
In the absence of lower bound check, the code fails later with ArrayIndexOutOfBoundException. The code doesn't break, however the user behavior is different. This PR unifies the exception user gets, when the index is not in range

### How was this patch tested?
Unit Test case was enhanced to add the test to validate the scenario

Closes #1616 from mystic-lama/dynamicIntArray_bound_checks.

Authored-by: mystic-lama <mysticlama000@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>