ORC-1506: Replacing deprecated valueOf() with recommended forNumber()
### What changes were proposed in this pull request?
The PR aims at replacing deprecated valueOf() with recommended forNumber().
valueOf() has been deprecated for a while and shall be removed in future versions. We are only removing it from the code we have written, it still exists in auto-generated code. Autogenerate code shall be compliant with version upgrade in future.

Couldn't find the exact PR for the change, however the reason seems to be performance related from search about the change

### Why are the changes needed?
The change helps protect against the removal of deprecated API in future releases.

### How was this patch tested?
No new tests were added, current tests pass

Closes #1619 from mystic-lama/deprecated_proto_valueOf.

Lead-authored-by: mystic-lama <mysticlama000@gmail.com>
Co-authored-by: Ash <mysticlama000@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>