ORC-1512: Drop Java 8/11 and make Java 17 by default
### What changes were proposed in this pull request?

This PR aims to drop Java 8 and 11 support and make Java 17 as the default and minimum Java version for Apache ORC 2.0.0.

This PR also upgrade GitHub Action test pipelines and Dockefiles.
- Update GitHub Action Java Versions to Java 17 and 21.
- Change GitHub Action Java Distribution to `Azul Zulu` from `Eclipse Temurin` because `temurin` doesn't have Java 21 yet.
- Update Dockerfiles to use Java 17 and 21
- Delete CentOS7 and Debian10 which requires non-trivial Java installations.

### Why are the changes needed?

Apache ORC 2.0 is aligned with Apache Spark 4.0 which dropped Java 8 and 11.
- https://issues.apache.org/jira/browse/SPARK-45315 (Drop JDK 8/11 and make JDK 17 by default)

In general, there are some extended support period for Java 11. However, Oracle's Premier Support ends this month, `September 2023`.
- https://www.oracle.com/java/technologies/java-se-support-roadmap.html

### How was this patch tested?

Pass the CIs

Closes #1627 from dongjoon-hyun/ORC-1512.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>