ORC-1614: Use `directOut.put(out)` instead of `directOut.put(out.array())` in `TestBrotli` test
### What changes were proposed in this pull request?
Set ByteBuffer limit in `TestBrotli` test

### Why are the changes needed?
`TestBrotli#testDirectDecompress` attempts to put the compressed result in direct ByteBuffer.
When calling `decompress`, we will find that the input buffer length is still `10000` not `217` since we put the array without `limit` on the length.

```java
      directOut.put(out.array());
      directOut.flip();
```

This PR is aimed to use `directOut.put(out)` instead of `directOut.put(out.array())`

### How was this patch tested?
GA

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #1790 from cxzl25/ORC-1614.

Authored-by: sychen <sychen@ctrip.com>
Signed-off-by: deshanxiao <deshanxiao@microsoft.com>