ORC-1618: Disable building tests for snappy
Disabling the building of tests is a workaround for issue #1791 and is probably not an issue since snappy tests are not executed as part of the build process of orc c++ libs.

Note that snappy 1.1.9 and 1.1.10 cannot be used as is because source assets on snappy's gh page are not buildable out of the box, so stick to 1.1.8 for now.

### What changes were proposed in this pull request?
Disable  building snappy tests as a workaround for issue #1791.

### Why are the changes needed?
compilation broken in some situations otherwise (see issue)

### How was this patch tested?
compile & unit tests

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #1792 from douardda/fix-snappy.

Authored-by: David Douard <david.douard@sdfa3.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit 9627a24439099046078c9f300028b79ab27fe33e)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>