ORC-1646: Close the reader when reading the schema with the `convert` command
### What changes were proposed in this pull request?
This PR aims to close the reader when reading the schema with the `convert` command.

### Why are the changes needed?
When using the `convert` command, when the input file format is ORC, the ORC schema is read, but not the close ORC reader.

### How was this patch tested?
local test

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #1837 from cxzl25/ORC-1646.

Authored-by: sychen <sychen@ctrip.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit 90dd4bd9fec80e391566e38a29cd02eaf4737af5)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>