ORC-1365: [C++] Use BlockBuffer to replace dataBuffer of  rawInputBuffer
### What changes were proposed in this pull request?
The purpose of this PR is to replace the DataBuffer with BlockBuffer. This way the orc compressor can start with a small initial size and grow automatically as needed.

### Why are the changes needed?
This patch uses BlockBuffer to replace the DataBuffer of class CompressionStream in order to solve the [issue](https://issues.apache.org/jira/browse/ORC-1365).

### How was this patch tested?
The UTs in TestBufferedOutputStream.cc and TestCompression.cc can cover this patch.

Closes #1916 from luffy-zh/ORC-1365.

Authored-by: luffy-zh <zhnice@outlook.com>
Signed-off-by: Gang Wu <ustcwg@gmail.com>