ORC-1742: Suppor print the id, name and type of each column in dump tool
### What changes were proposed in this pull request?
This PR aims to suppor print the id, name and type of each column in dump tool.

### Why are the changes needed?
When we dump an ORC with a complex structure, we only output the column id of each column, but we do not know the column name and type corresponding to the column id.
If we use json format, it will output the id, name and type of each column.

### How was this patch tested?
Add UT

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #1974 from cxzl25/ORC-1742.

Authored-by: sychen <sychen@ctrip.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit 8ca3a23975d9018ea3311ba11431303ae091ec5a)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>