ORC-1745: Remove `Ubuntu 20.04` Support
### What changes were proposed in this pull request?

This PR aims to remove `Ubuntu 20.04` Support from Apache ORC 2.1.0.

1. Remove from GitHub CIs
2. Remove from `docker` directory
3. Remove from Apache ORC Website.

### Why are the changes needed?

Apache ORC 2.1.0 is scheduled in 2025 and Ubuntu 20.04 LTS is going to reach the end of support April 2025. We had better focus on the latest OSes at Apache ORC 2.1.0.

| Version | RELEASED | END OF STANDARD SUPPORT |
| - | - | - |
| Ubuntu 20.04 LTS (Focal Fossa) | Apr 2020 | Apr 2025 |

### How was this patch tested?

Pass the CIs. However, it will pass because this is a removal of test coverage.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #1983 from dongjoon-hyun/ORC-1745.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>