MINOR: Fix syntax of DOAP
### What changes were proposed in this pull request?

I'm having some trouble finding an authoritative reference, but it looks like a release (which is a property) may not contain multiple Versions (which is a node) in RDF/XML. The python rdflib at least doesn't support it, and whatever generates https://projects.apache.org/json/projects/orc.json and https://projects.apache.org/project.html?orc also doesn't seem to process it correctly.

### Why are the changes needed?

So infrastructure that uses the information for the DOAP does the right thing.

### How was this patch tested?

n/a

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #1964 from raboof/patch-1.

Authored-by: Arnout Engelen <arnout@bzzt.net>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>