Fix supportVectoredIO for bad hadoop version string
There are cases where the hadoop version info may not be respecting
the semantic versioning. It is the case for the hadoop version
provided in some of the AWS managed services. This causes
a ExceptionInInitializerError while trying to instantiate and ORC
file reader.

Fixing this by defaulting to non vectored IO in case the semantic
versioning is not respected