ORC-1749: Fix `supportVectoredIO` for hadoop version string with optional patch labels
### What changes were proposed in this pull request?
Parse correctly semantic versioning when there is the optional labels in the patch are present.

### Why are the changes needed?
There are cases where the hadoop version info may not be respecting the semantic versioning. It is the case for the hadoop version provided in some of the AWS managed services. This causes a ExceptionInInitializerError while trying to instantiate and ORC file reader.

### How was this patch tested?
Included unit tests. It required a small refactor to be able to test it in an easier way

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #1990 from otorreno/fix/supportVectoredIO.

Authored-by: Oscar Torreno <oscartorreno@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>