ORC-1683: Fix `instanceof` of BinaryStatisticsImpl merge method
### What changes were proposed in this pull request?
This PR aims to fix `instanceof` of `BinaryStatisticsImpl` merge method.

### Why are the changes needed?
In [ORC-1542](https://issues.apache.org/jira/browse/ORC-1542), we modified part of instanceof, but `BinaryStatisticsImpl` was not modified because the merge method was written differently.

However, the current code is instanceof `BinaryColumnStatistics` and then explicitly cast `BinaryStatisticsImpl`, so it should be replaced by the new instanceof writing method (`Pattern Matching for instanceof`).

```java
if (other instanceof BinaryColumnStatistics) {
        BinaryStatisticsImpl bin = (BinaryStatisticsImpl) other;
```

### How was this patch tested?
GA

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #1883 from cxzl25/ORC-1683.

Authored-by: sychen <sychen@ctrip.com>
Signed-off-by: Shaoyun Chen <csy@apache.org>