ORC-1264: [C++] Add a writer option to align compression block with row group
## What changes were proposed in this pull request?
Add support for the ORC writer to ensure that the compression block is aligned with the row group boundary。

## Why are the changes needed?
To reduce unnecessary I/O and decompression when PPD is in effect, we can enforce the compression block to be aligned with the row group boundary. For more detail, see [link](https://issues.apache.org/jira/projects/ORC/issues/ORC-1264?filter=allopenissues)

## How was this patch tested?
Uts in TestWriter.cc can convert this patch.

## Was this patch authored or co-authored using generative AI tooling?
NO

Closes #2005 from luffy-zh/ORC-1264.

Lead-authored-by: luffy-zh <zhnice@outlook.com>
Co-authored-by: Hao Zou <34559830+luffy-zh@users.noreply.github.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>