ORC-1880: [C++] Add invalid argument check for NOT Operator in ExpressionTree
### What changes were proposed in this pull request?

Add invalid argument check for NOT Operator in ExpressionTree.cc

### Why are the changes needed?

Operator::NOT expects to have exactly one child, but there is currently no validation to enforce this requirement. If no child is provided, it may result in a crash during the construction of the SearchArgument.

### How was this patch tested?

The unit tests in TestSearchArgument.cc can effectively cover this patch.

### Was this patch authored or co-authored using generative AI tooling?

NO

Closes #2196 from luffy-zh/ORC-1880.

Authored-by: luffy-zh <zhnice@outlook.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>