ORC-1933: Change `org.jetbrains:annotations` dependency to the `provided` scope
### What changes were proposed in this pull request?

This PR aims to lower `org.jetbrains:annotations` dependency from `compile` scope to `provided` scope for Apache ORC 2.2.0.

From now, we can update it to the latest version because this doesn't affect the downstream. So, this PR also
- Remove the pinning the dependency.
- Upgrade it to the latest version

### Why are the changes needed?

To simplify Apache ORC dependency because this is not required at the runtime.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #2293 from dongjoon-hyun/ORC-1933.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>