ORC-1937: Use the default `buildtype` in Meson config
### What changes were proposed in this pull request?
This removes setting the Meson configuration to remove the buildtype of release by default.

### Why are the changes needed?
While coming from CMake it may seem like a good idea to build release builds by default, the setting is either ignored or problematic when orc is used as a subproject in other projects, especially on Windows where a main project building a library with debug symbols may fail if a subproject tries to force release symbols.

### How was this patch tested?
Compiled locally

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #2298 from WillAyd/remove-meson-release-build.

Authored-by: Will Ayd <william.ayd@icloud.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>