ORC-1884: [C++] Add a maybe() API to the SearchArgumentBuilder
### What changes were proposed in this pull request?

Add a maybe() API to the SearchArgumentBuilder to allow the insertion of an expression tree that is currently unsupported.

### Why are the changes needed?

To enhance the SearchArgumentBuilder API with a new method called maybe(), we can create a SearchArgument tree for scenarios that may not be adequately addressed by existing methods.

### How was this patch tested?

Uts in TestSearchArgument.cc can cover this patch.

### Was this patch authored or co-authored using generative AI tooling?

NO

Closes #2207 from luffy-zh/ORC-1884.

Authored-by: luffy-zh <zhnice@outlook.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>